==> R/00_doit.R <==
# Run all scripts ---------------------------------------------------------

source(file = "R/01_load.R")
source(file = "R/02_clean.R")
source(file = "R/03_augment.R")
source(file = "R/04_analysis_i.R")
source(file = "R/05_analysis_ii.R")
source(file = "R/06_PCA.R")
